requirements.txt
# Dependencies untuk Google Scholar Scraper

# Web Scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# File Handling
openpyxl>=3.1.0
python-docx>=1.0.0

# Utilities
requests>=2.31.0
